refactor: move AtmosphereModel class and commented classes ExponentialAtmosphere, SimpleExponentialAtmosphere to new atmosphere_model.py
