+++ gmatpyplus/solar_system/__init__.py
@@ -1,0 +1,1 @@
+from .atmosphere_model import *
